…factor

- [x] *The Jira issue number for this PR is: MDEV-26356*

<!--
An amazing description should answer some questions like:
1. What problem is the patch trying to solve?
2. If some output changed, what was it looking like before
   the change and how it's looking with this patch applied
3. Do you think this patch might introduce side-effects in
   other parts of the server?
-->
## Description
Users can configure the number of purge threads to use.

Currently, the existing logic looks at the increasing history length and
accordingly increases the purge threads (respecting the user-set threshold).
User active user workload causes history length to increase that, in turn,
leads to more purge threads to get scheduled.

Purge operation generates redo-records in addition to user workload
redo-records. This dual redo generations cause intense pressure on redo-log
which can easily cause redo-log to hit the threshold and there-by-causing
jitter in overall throughput (including furious flushing).

In order to improve the user workload performance and it is important to ensure
background activity like a purge is not overdone while ensuring history length
is kept in check to avoid an increase in select latency.

The proposed patch explores an adaptive purge thread scheduling approach based
on the redo log fill factor. Logic tends to dynamically increase or decrease
the purge threads based on how much redo log is filled, leaving enough
space/resources/flush bandwidth for the user workload.

## How can this PR be tested?
This patch is about performance. To test it user can run CPU-Bound workload with and without patch and observe the effect especially when slower SSD (say throughput of 350 MB/sec).  For more configuration details: https://github.com/mysqlonarm/benchmark-suites/blob/master/mysql-sbench/conf/mdb.cnf/100tx1.5m_cpubound.cnf.
Users can play around with innodb_purge_threads. Normally, on a slower disk, the default purge thread of 4 is enough to produce a performance jitter.

<!--
Tick one of the following boxes [x] to help us understand
if the base branch for the PR is correct
-->
## Basing the PR against the correct MariaDB version
- [x ] *This is a new feature and the PR is based against the latest MariaDB development branch*
- [ ] *This is a bug fix and the PR is based against the earliest branch in which the bug can be reproduced*

<!--
You might consider answering some questions like:
1. Does this affect the on-disk format used by MariaDB?
2. Does this change any behavior experienced by a user
   who upgrades from a version prior to this patch?
3. Would a user be able to start MariaDB on a datadir
   created prior to your fix?
-->
## Backward compatibility
1. Doesn't affect the on-disk format
2. User should see improved performance. Dynamic adjustment of purge_threads is an existing logic we improved it to use redo-log factor so no functional difference there but depending on redo log user may see increase in history length (even possible w/o patch) there-by more disk space usage.
3. Yes, the User can start the server using the existing data directory.