@dr-m 2021-09-02
I think that larger changes around the purge subsystem are necessary. There are some nice ideas in [MDEV-16260](https://jira.mariadb.org/browse/MDEV-16260).

I just applied a rather dummy work-around 2224e15e425db7568df7b0e10a3a9b793c05ea6e to thwart a counter-intuitive performance regression that we experienced due to [MDEV-24258](https://jira.mariadb.org/browse/MDEV-24258). It might be worthwhile to test another type of work-around of acquiring and releasing `table->lock_mutex`, to validate or reject the following idea.

What if we maintained something like `std::unordered_map<table_id_t,worker>` so that all entries for a particular table or transaction would always be assigned to the same purge task, to minimize contention elsewhere between purge tasks. As soon as all currently buffered work is completed for the table, we could delete the mapping. Another idea would be to assign all undo log of a particular transaction to the same worker task.

One more idea: What if the workload does not fit in the buffer pool? In addition to monitoring the checkpoint age, should we also monitor the page eviction rate that is caused by purge tasks?