@mysqlonarm 2021-09-03
1. table-id based purge allocation could be considered. Infact, MySQL has it and lately they hit this skewed allocation issue that they fixed. (Bug #32089028). Better done a independent functional change.
2. I didn't hit the issue of purge contention with IO bound workload. During your experiment did you hit the issue of purge contention for IO bound workload that could be eased by taking into consideration page eviction rate. Please share the use-case and I will try. This if needed could be considered as part of the said ticket or PR.