@dr-m 2021-09-03
I believe that the regression that 2224e15e425db7568df7b0e10a3a9b793c05ea6e tries to alleviate occurred because the purge coordinator and purge workers used to spend most of their waiting time on `dict_sys.mutex` or `dict_sys.latch` and some time on a page latch (undo page latch in the purge coordinator, and index tree page latch in the workers). After `dict_sys.mutex` was removed, all of the waits would be on page latches. This could mean that purge is causing too much pressure on the buffer pool. I do not have the configuration parameters of that test run; I only saw some http://www.brendangregg.com/offcpuanalysis.html graphs. The scenario might involve a buffer pool that is ¾ or ½ the size of the working set, so that some LRU eviction would kick in.

Note: I tested the impact of 2224e15e425db7568df7b0e10a3a9b793c05ea6e both with a CPU-bound configuration (everything fits in the buffer pool) and a configuration where LRU eviction should take place. In all configurations that I tested (using the default `innodb_purge_threads=4`), the throttling seemed to help. Of course, it is only a workaround, which hopefully can be removed as part of this pull request.

Can we add the following inputs to the adaptive algorithm?
1. The redo log volume (LSN growth) caused by purge
1. The number of `buf_pool_t::insert_into_flush_list()` calls by purge

If the server is running a heavy write workload, we might want to limit the resource consumption of purge to some fraction of all writes.

One more somewhat related idea is in [MDEV-26475](https://jira.mariadb.org/browse/MDEV-26475), to run part of the purge in user threads that are about to acquire locks on a page, or modifying a page. That could better combine writes to index pages. A purge worker might then only notice ‘oh, the record was already purged’ and move on. In an extreme case of this, only the purge coordinator would perform writes, when discarding the undo pages (truncating the history).