@dr-m 2021-09-13
>     1. table-id based purge allocation could be considered. Infact, MySQL has it and lately they hit this skewed allocation issue that they fixed. (Bug #32089028). Better done a independent functional change.

This is interesting, and I agree. I see that there were multiple follow-up commits to mysql/mysql-server@5ba03cadf63b6c53a780ffd887d903eae6aac4ee so far.

>     2. I didn't hit the issue of purge contention with IO bound workload. During your experiment did you hit the issue of purge contention for IO bound workload that could be eased by taking into consideration page eviction rate. Please share the use-case and I will try. This if needed could be considered as part of the said ticket or PR.

I hope that @hgxl64 can share what exactly he did. For what it is worth, I just tested `oltp_update_non_index` on 100×1.5M rows (about 40GiB of data) with `innodb_buffer_pool_size=30G`. I compared 10.6 329d370f2d6af13cee89bc2c8db0f5915dcbfd4a with a revert of the hack 4690442411b75ea0fc1a6aacabe767d6fb1d1f62 combined to a cherry-pick of the latest commit of this pull request (CommitDate: August 25).
|threads|10.6/tps|patched/tps|
|------:|-----:|----:|
|32|240486.86|234973.63|
|64|222660.59|224189.97|
|128|213950.89|218025.55|

The benchmark was run for 120 seconds for each thread count. At the end, there was a shutdown with `innodb_fast_shutdown=0` and `innodb_buffer_pool_dump_pct=100`. At the start, there was a `sleep` to let the buffer pool to be loaded before starting workload. A longer benchmark would be needed to truly assess the impact. Because the buffer pool is smaller than the data files, the results are not fully deterministic or repeatable.

In any case, I will try to implement my suggestion to monitor the `buf_pool_t::insert_into_flush_list()` calls by purge.