@dr-m 2021-09-14
I created my own version of this, which I hoped to be roughly equivalent to this. I did not try to implement my idea of taking into account the redo log volume caused by purge tasks, or the `buf_pool_t::insert_into_flush_list()` calls invoked by purge tasks.
I repeated my above benchmark for 3 commits in a 10.6-based branch:
|name|commit|
|---|---|
|baseline|03e4cb2484cc302bef2886dd3d76b3acdf7f5c23|
|refactored|717a32154cc6de81d1a0d3a820928b78aa7b1140|
|adaptive|ea52a3eb9795f1ae04a4f9c2be4c81e43dc86c15|

These were built with `cmake -DPLUGIN_PERFSCHEMA=NO`. The performance at 32 threads is suspiciously low, compared to my earlier test, but it could be explained by an operating system upgrade or by concurrent use of a graphical user environment on the system.
|threads|baseline/tps|refactored/tps|adaptive/tps|
|----:|----:|----:|----:|
|32|215146.80|199784.55|204049.46|
|64|219921.36|221165.78|216341.43|
|128|183037.61|191960.11|188649.15|

The refactoring simplifies the scheduling of the purge coordinator task. Based on the above very quick benchmark, it might help at larger thread counts and hurt performance at lower thread counts. Maybe we should simply make the purge somehow more aggressive when there is not too much concurrent user workload?

More extensive tests are needed, including monitoring of the log checkpoint age, page write rate, LRU eviction rate, and the history list length.